requirements.txt
cbor2==5.4.6
certifi==2022.12.7
influxdb-client==1.36.1
luma.core==2.4.0
luma.emulator==1.4.0
luma.lcd==2.10.0
pigpio==1.78
Pillow==9.5.0
pyftdi==0.54.0
pygame==2.3.0
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
reactivex==4.0.4
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.4.2
spidev==3.6
typing-extensions==4.5.0
urllib3==1.26.15
